refactor: move PlaySVG from atoms to molecules
