+++ src/components/molecules/PlaySVG/index.vue
