+++ src/components/organisms/SideController/index.vue
@@ -22,8 +22,8 @@
 </template>
 
 <script>
-import PlaySVG from "@/components/atoms/PlaySVG";
 import Logo from "@/components/atoms/Logo";
+import PlaySVG from "@/components/molecules/PlaySVG";
 
 export default {
   name: "sideController",
